+++ UPM/Runtime/Bootstrap/Bootstrap.cs
@@ -2,11 +2,18 @@
 using System.Linq;
 using UnityEngine;
 
-namespace E314.GameKit.Uni
+namespace E314.ExtUni
 {
 
+/// <summary>
+/// Abstract class for managing application initialization.
+/// Provides basic functionality for loading and initializing system components.
+/// </summary>
 public abstract class Bootstrap : ScriptableObject
 {
+	/// <summary>
+	/// Gets the singleton instance of Bootstrap.
+	/// </summary>
 	internal static Bootstrap Instance { get; private set; }
 
 #if UNITY_EDITOR
@@ -16,6 +23,11 @@ public abstract class Bootstrap : ScriptableObject
 	}
 #endif
 
+	/// <summary>
+	/// Initializes components depending on the loading stage.
+	/// </summary>
+	/// <param name="type">Type of initialization stage.</param>
+	/// <exception cref="ArgumentOutOfRangeException">Thrown when initialization type is unknown.</exception>
 	internal void Initialize(RuntimeInitializeLoadType type)
 	{
 		switch (type)
@@ -40,14 +52,29 @@ public abstract class Bootstrap : ScriptableObject
 		}
 	}
 
+	/// <summary>
+	/// Called during subsystem registration.
+	/// </summary>
 	protected abstract void OnSubsystemRegistration();
 
+	/// <summary>
+	/// Called after all assemblies are loaded.
+	/// </summary>
 	protected abstract void OnAfterAssembliesLoaded();
 
+	/// <summary>
+	/// Called before the splash screen is displayed.
+	/// </summary>
 	protected abstract void OnBeforeSplashScreen();
 
+	/// <summary>
+	/// Called before scene loading.
+	/// </summary>
 	protected abstract void OnBeforeSceneLoad();
 
+	/// <summary>
+	/// Called after scene loading.
+	/// </summary>
 	protected abstract void OnAfterSceneLoad();
 
 	private void Awake()