+++ UPM/CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0] - 2025.03.09
 
 ### Added
 
-- 
+- Bootstrap: Provides basic functionality for loading and initializing system components.
+- SerializableType: Allows serialization of System.Type for Unity inspector. Displays as a dropdown list in Unity inspector.